Fix some format string errors
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9534 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/smbios.cpp
@@ -227,7 +227,7 @@ void FieldInitializer::operator()<const char*>(size_t flags, const char*& t, con
 
 	if(number > strings.size())
 	{
-		debug_printf(L"SMBIOS: invalid string number %d (count=%d)\n", number, strings.size());
+		debug_printf(L"SMBIOS: invalid string number %d (count=%d)\n", number, (int)strings.size());
 		return;
 	}
